requirements.txt
asgiref==3.2.3
astroid==2.3.3
Click==7.0
Django==3.0.1
django-secrets==1.0.2
django-tastypie==0.14.2
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pur==5.3.0
pylint==2.4.4
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wrapt==1.11.2
